refactored where device info methods are
